Launch and Target
We added a launch method and a cooldown method. We also added variables to track the amount of time passed from the start and end of each launch. I made the pin the target for all the balls and added a homing missel behavior for the balls to follow the pin.