Score Board
I added a Score panel that adds a point every time you successfully dodge a launched ball. I created a new script for it that updates the text. I varied the launching duration for the balls to be different so that they are not all launching around the same time.